Refactor the code
- Save the hyperscaler API jsons
- Save the docling document formats
- Temporary files are stored as below:
   - `intermediate_files` -- for parquet files
   - `microsoft` - Root folder that contains API output files
      - `docling_document` - docling document output of the MS outputs
   - `visualizations` - Output of table visualizations
