+++ docling_eval/benchmarks/utils.py
@@ -390,8 +390,6 @@ def save_shard_to_disk(
     batch = Dataset.from_list(items)  # , features=features)
 
     output_file = dataset_path / f"shard_{thread_id:06}_{shard_id:06}.{shard_format}"
-    logging.info(f"Saved shard {shard_id} to {output_file} with {len(items)} documents")
-
     if shard_format == "json":
         batch.to_json(output_file)
 
@@ -400,6 +398,8 @@ def save_shard_to_disk(
 
     else:
         raise ValueError(f"Unsupported shard_format: {shard_format}")
+
+    logging.info(f"Saved shard {shard_id} to {output_file} with {len(items)} documents")
 
     shard_id += 1
